+++ src/settings/AppearanceSettings.java
@@ -2,17 +2,8 @@ package settings;
 
 public class AppearanceSettings {
 
-    public static final String WHITE = "\033[0;38m";
-    public static final String RESET = "\033[0m";
-    public static final String RED = "\033[0;31m";
-    public static final String GREEN = "\033[0;32m";
-    public static final String YELLOW = "\033[0;33m";
-    public static final String BLUE = "\033[0;34m";
-    public static final String PURPLE = "\033[0;35m";
-    public static final String CYAN = "\033[0;36m";
+    //IN PROGRESS
 
-    public static final String BOLD = "\033[1m";
-    public static final String UNDERLINE = "\033[4m";
     //public static final String GRAY = "\033[0;37m";
 }
    /* static Scanner scanner = new Scanner(System.in);
